refactored datainit logic
