--- src/main/java/adeuxpas/back/datainit/ArticlePictureSeeder.java
@@ -1,4 +1,0 @@
-package adeuxpas.back.datainit;
-
-public class ArticlePictureSeeder {
-}
