--- src/main/java/adeuxpas/back/datainit/UserSeeder.java
@@ -1,4 +1,0 @@
-package adeuxpas.back.datainit;
-
-public class UserSeeder {
-}
